+++ base/schema/index.js
@@ -5,30 +5,43 @@ const fn = require('zerodep/node/fn');
 module.exports = (...dirPathResolveArgs) => {
     const files = require('zerodep/node/tree/json')(...dirPathResolveArgs);
     const schema = {};
+    // extract schemas
+    const extractSchema = (...keys) => {
+        const target = fn.get(files, ...keys);
+        let id;
+        if (typeof target.id === 'string') id = target.id;
+        if (typeof target.$id === 'string') id = target.$id;
+        // extract base schema and nested schemas
+        if (id) (schema[id] = target) && extractNestedSchema(target, id);
+    };
+    // extract nested schemas
+    const extractNestedSchema = (target, base) => {
+        fn.assignDeepKeyParent(
+            /^(\$id|id)$/,
+            (target, parentKey, key) => {
+                if (typeof target[key] === 'string') {
+                    const id = new URL(target[key], base).href.replaceAll('##', '#');
+                    schema[id] = target;
+                    if (/^(\$id|id)$/.test(...fn.deepKeys(target))) extractNestedSchema(target, id);
+                    return { [parentKey]: { $ref: id } };
+                }
+            },
+            target
+        );
+    };
     // all individual schemas will run through this
     const decentralize = (id) => {
         // re-reference
         fn.replaceDeepKey(
             '$ref',
             (ref) => {
-                if (typeof ref === 'string') {
-                    let url;
-                    if (ref.indexOf('http') === 0) {
-                        url = new URL(`${ref}`);
-                    } else if (ref.indexOf('#') === 0) {
-                        url = new URL(`${id}${ref}`);
-                    } else {
-                        url = new URL(`${id.split('/').slice(0, -1).join('/')}/${ref}`);
-                    }
-                    return [{ ref: url.href.replace('##', '#') }];
-                } else {
-                    return [{ ref }];
-                }
+                if (typeof ref === 'string') return { ref: new URL(ref, id).href.replaceAll('##', '#') };
+                else return { ref };
             },
             schema[id]
         );
         fn.replaceDeepKey(
-            '$recursiveRef',
+            /\$recursiveRef|\$dynamicRef/,
             (ref) => {
                 if (typeof ref === 'string') {
                     let url;
@@ -43,30 +56,8 @@ module.exports = (...dirPathResolveArgs) => {
                     } else {
                         url = new URL(`${id}/----unexpected-case----${ref}`);
                     }
-                    return [{ ref: url.href.replace('##', '#') }];
-                } else {
-                    return []; // remove definition object
-                }
-            },
-            schema[id]
-        );
-        fn.replaceDeepKey(
-            '$dynamicRef',
-            (ref) => {
-                if (typeof ref === 'string') {
-                    let url;
-                    if (ref.indexOf('http') === 0) {
-                        if (ref.indexOf('#') > 0) {
-                            url = new URL(`${ref.substring(0, ref.indexOf('#'))}`);
-                        } else {
-                            url = new URL(`${ref}`);
-                        }
-                    } else if (ref.indexOf('#') === 0) {
-                        url = new URL(`${id}`);
-                    } else {
-                        url = new URL(`${id}/----unexpected-case----${ref}`);
-                    }
-                    return [{ ref: url.href.replace('##', '#') }];
+                    return { ref: url.href.replace('##', '#') };
+                    // return { ref: new URL(ref, id).href.replace('##', '#') };
                 } else {
                     return []; // remove definition object
                 }
@@ -90,15 +81,9 @@ module.exports = (...dirPathResolveArgs) => {
         fn.replaceDeepKey('definitions', () => [], schema[id]);
         fn.replaceDeepKey('$defs', () => [], schema[id]);
     };
-    // id parser
-    const parser = (...keys) => {
-        const target = fn.get(files, ...keys);
-        if (typeof target.id === 'string') (schema[target.id] = target) && decentralize(target.id);
-        if (typeof target.$id === 'string') (schema[target.$id] = target) && decentralize(target.$id);
-    };
-    // extract id or $id
-    fn.parseDeepKeyParent('id', parser, files);
-    fn.parseDeepKeyParent('$id', parser, files);
+    // extract individual schemas (base schemas, then schemas nested inside of them)
+    fn.parseDeepKeyParent(/^(\$id|id)$/, extractSchema, files);
+    // decentralize schemas
+    Object.keys(schema).forEach((id) => decentralize(id));
     return schema;
 };
-