+++ config/env/index.js
@@ -1,0 +1,9 @@
+'use strict';
+const dotenv = require('dotenv');
+const result = dotenv.config({ silent: true, path: `./config/env/.env` });
+
+// test config
+if (result.error) throw result.error;
+// console.log('NODE_ENV:', result.parsed);
+
+module.exports = dotenv;